requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyJWT==2.8.0
bcrypt==4.0.1
google-auth==2.22.0
python-dotenv==1.0.0
openai==0.28.1
pymongo==4.7.2
dnspython==2.4.2